đã hoàn thành chức năng upload file, còn file key thì sẽ gửi riêng để cấu hình sau
