+++ .gitignore
@@ -31,3 +31,6 @@ build/
 
 ### VS Code ###
 .vscode/
+
+### bỏ qua commit file chứa key upload file gg để tránh lộ key ###
+src/main/resources/credentials/drive-service-account.json
