+++ pom.xml
@@ -111,6 +111,25 @@
 			<version>2.0.0</version>
 		</dependency>
 
+		<dependency>
+			<groupId>com.google.oauth-client</groupId>
+			<artifactId>google-oauth-client</artifactId>
+			<version>1.34.1</version>
+		</dependency>
+
+		<dependency>
+			<groupId>com.google.http-client</groupId>
+			<artifactId>google-http-client</artifactId>
+			<version>1.40.1</version> <!-- hoặc 1.40.1 nếu cần mới hơn -->
+		</dependency>
+
+		<dependency>
+			<groupId>com.google.oauth-client</groupId>
+			<artifactId>google-oauth-client-jetty</artifactId>
+			<version>1.31.5</version>
+		</dependency>
+
+
 
 		<!-- Google API Client jackson -->
 		<dependency>
