+++ src/main/java/com/codebloom/cineman/config/GoogleDriveConfig.java
@@ -1,40 +1,53 @@
 package com.codebloom.cineman.config;
-
+import com.google.api.client.googleapis.auth.oauth2.GoogleAuthorizationCodeFlow;
+import com.google.api.client.googleapis.auth.oauth2.GoogleClientSecrets;
 import com.google.api.client.googleapis.javanet.GoogleNetHttpTransport;
+import com.google.api.client.extensions.java6.auth.oauth2.AuthorizationCodeInstalledApp;
+import com.google.api.client.extensions.jetty.auth.oauth2.LocalServerReceiver;
+import com.google.api.client.json.JsonFactory;
 import com.google.api.client.json.jackson2.JacksonFactory;
+import com.google.api.client.util.store.FileDataStoreFactory;
 import com.google.api.services.drive.Drive;
 import com.google.api.services.drive.DriveScopes;
-import com.google.auth.http.HttpCredentialsAdapter;
-import com.google.auth.oauth2.GoogleCredentials;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.core.io.ClassPathResource;
-import org.springframework.core.io.Resource;
 
-import java.io.IOException;
+import java.io.InputStreamReader;
 import java.io.InputStream;
-import java.security.GeneralSecurityException;
-import java.util.Collections;
+import java.nio.file.Paths;
 import java.util.List;
 
 @Configuration
 public class GoogleDriveConfig {
-    @Value("${google.service.account.key}")
-    private Resource accountKey;
+
+    private static final JsonFactory JSON_FACTORY = JacksonFactory.getDefaultInstance();
+    private static final List<String> SCOPES = List.of(DriveScopes.DRIVE_FILE);
 
     @Bean
-    public Drive googleDrive() throws IOException, GeneralSecurityException {
-        InputStream in = new ClassPathResource("credentials/drive-service-account.json").getInputStream();
+    public Drive googleDrive() throws Exception {
+        InputStream in = getClass().getClassLoader().getResourceAsStream("credentials.json");
+        if (in == null) throw new RuntimeException("credentials.json not found");
 
-        GoogleCredentials credentials = GoogleCredentials.fromStream(in)
-                .createScoped(List.of(DriveScopes.DRIVE));
+        GoogleClientSecrets clientSecrets = GoogleClientSecrets.load(JSON_FACTORY, new InputStreamReader(in));
+
+        var flow = new GoogleAuthorizationCodeFlow.Builder(
+                GoogleNetHttpTransport.newTrustedTransport(),
+                JSON_FACTORY,
+                clientSecrets,
+                SCOPES
+        )
+                .setDataStoreFactory(new FileDataStoreFactory(Paths.get("tokens").toFile()))
+                .setAccessType("offline")
+                .build();
+
+        var receiver = new LocalServerReceiver.Builder().setPort(8888).build();
+
+        var credential = new AuthorizationCodeInstalledApp(flow, receiver).authorize("user");
 
         return new Drive.Builder(
                 GoogleNetHttpTransport.newTrustedTransport(),
-                JacksonFactory.getDefaultInstance(),
-                new HttpCredentialsAdapter(credentials))
-                .setApplicationName("cineman")
-                .build();
+                JSON_FACTORY,
+                credential
+        ).setApplicationName("cineman").build();
     }
 }
